Upgrade Laundry Light to use ComplexLight and fix ComplexLight
Added a motion sensor for laundry light, so it can use complexlight.

complexlight was incorrectly calling movementlight and was also not paying attention to the switch the right way.

Now it will turn on the light without any delay and won't kick off the timeout, if the motion sensor is then triggered, the timeout will be started.

Boo@ bosch motion sensors only "detecting motion" for 4 seconds every 3 minutes.