[Calendar Switch] Adding debugging for token expiration
Summary: It seems that the tokens will eventaully expire and be unable
be renewed automatically. I'm thinking this will be probably a weekly
occurrence (since it seems to have happend ~1 week since I created the
setup).

Will test if it is, and if so, continue looking into how to possibly get
a token that can actually last indefinitely, or maybe I need to change
the way we're querying the API so I don't need to continually grant
acess to the app (could dig into what HA is doing with the TV stuff).

Test Plan: Deleted the token file to simulate a bad token, debug speaker
properly announced failure on app restart