[Calendar Switch] Fix auth setup to use home assistant configured token
Summary: After a bit of futzing, the issue was made apparent that the problem
with using the token from the setup within home assistant was because
home assitant was using the base scope, and the quickstart example was
using the readonly scope. Apparently the base scope does not grant
access to the readonly scope, so we can just request the base scope and
get be able to use the same token. Since this token was set up using the
"TV and other devices" using a pairing code, I expect it to last
indefinitely. We'll see if that's true next week since it does appear to
be expiring at exactly 1 week.

Test plan: Pushed to the pi, saw that it successfully retrieved events
using the configured token.