Create RGB Clock
Creating the RGB clock. Use any RGB light as an analog clock by having the hue rotate 360 degrees every 12 hours, so you can easily see what time it is based on the color.  Imagine a color wheel, and the hour hand points to what color will be displayed (e.g. 12:00 is red, 3:00 is green, 6:00 is teal, 9:00 is purple). Additionally, the clock will blink the number of the hour at the bottom of the hour every hour, and go dark for 2 seconds at each half hour mark.

Test plan:
- Checked that correct time is shown for current time
- Hardcoded values to check for various situations:
  - Check that *:30 correctly shows half hour animation
  - Check that *:00 correctly shows various time animations